added database to few more user routes

+++ src/model/user_model.rs
@@ -6,7 +6,7 @@ pub struct User {
     #[serde(rename = "_id", skip_serializing_if = "Option::is_none")]
     pub id: Option<ObjectId>,
     pub name: String,
-    pub unqid: u64,
+    pub unqid: u32,
     pub image: String,
     pub rooms: Vec<ObjectId>,
     pub email: String,
